api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.3.12",
    "anthropic>=0.89.0",
    "fastapi>=0.124.4",
    "fugashi>=1.5.2",
    "google-genai>=1.55.0",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "neo4j>=6.0.3",
    "openpyxl>=3.1.5",
    "pdfplumber>=0.11.8",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pykakasi>=2.3.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "unidic-lite>=1.0.8",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
